refactor: 정의 클래스 글로벌로 이동
